+++ core/Updates/4.0.0-b1.php
@@ -9,6 +9,8 @@
 
 namespace Piwik\Updates;
 
+use Piwik\Date;
+use Piwik\Plugins\UsersManager\Model;
 use Piwik\Updater;
 use Piwik\Updates as PiwikUpdates;
 use Piwik\Updater\Migration;
@@ -31,13 +33,33 @@ class Updates_4_0_0_b1 extends PiwikUpdates
 
     public function getMigrations(Updater $updater)
     {
-        $migration1 = $this->migration->db->changeColumnType('log_action', 'name', 'VARCHAR(4096)');
-        $migration2 = $this->migration->db->changeColumnType('log_conversion', 'url', 'VARCHAR(4096)');
+        $migrations = array();
+        $migrations[] = $this->migration->db->changeColumnType('log_action', 'name', 'VARCHAR(4096)');
+        $migrations[] = $this->migration->db->changeColumnType('log_conversion', 'url', 'VARCHAR(4096)');
 
-        return array(
-            $migration1,
-            $migration2
-        );
+        $migrations[] = $this->migration->db->createTable('user_token_auth', array(
+            'idusertokenauth' => 'BIGINT UNSIGNED NOT NULL AUTO_INCREMENT',
+            'login' => 'VARCHAR(100) NOT NULL',
+            'description' => 'VARCHAR('.Model::MAX_LENGTH_TOKEN_DESCRIPTION.') NOT NULL',
+            'password' => 'VARCHAR(255) NOT NULL',
+            'last_used' => 'DATETIME NULL',
+            'date_created' => ' DATETIME NOT NULL',
+        ), 'idusertokenauth');
+        $migrations[] = $this->migration->db->addUniqueKey('user_token_auth', 'password', 'uniq_password');
+
+        $userModel = new Model();
+        foreach ($userModel->getUsers(array()) as $user) {
+            $migrations[] = $this->migration->db->insert('user_token_auth', array(
+                'login' => $user['login'],
+                'description' => 'Created by Matomo 4 migration',
+                'password' => $userModel->hashTokenAuth($user['token_auth']),
+                'date_created' => Date::now()->getDatetime()
+            ));
+        }
+
+        $migrations[] = $this->migration->db->dropColumn('user', 'token_auth');
+
+        return $migrations;
     }
 
     public function doUpdate(Updater $updater)